updated menuState, Added Sprite object for Menu, Added Helpers class
I created a package sprites  with Classes for each Sprite of a state. Each classes extends of Sprite.

I created a Helpers class. I have 2 Methods to have real proportion of a dimension given (width or height) compared to the mobile resolution. For the reference, I chose 1440 x 2260 (max for me), so all assets images are created for this resolution and I use these two methods to adapt the size for the current mobile resolution.